separated PostService and PostFetchService responsibilities
