+++ src/main/java/pl/apzumi/task/service/PostFetchService.java
@@ -3,16 +3,11 @@ package pl.apzumi.task.service;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
-import pl.apzumi.task.domain.PostEntity;
 import pl.apzumi.task.dto.PostDto;
-import pl.apzumi.task.mappers.PostMapper;
-import pl.apzumi.task.repository.PostRepository;
 
-import javax.annotation.PostConstruct;
 import java.net.ConnectException;
 import java.util.Arrays;
 import java.util.List;
@@ -24,23 +19,8 @@ import java.util.stream.Collectors;
 @Transactional
 public class PostFetchService {
     public static final String BASE_API_URL = "https://jsonplaceholder.typicode.com/posts";
-    private final PostRepository postRepository;
-    private final PostMapper postMapper;
 
-    @Scheduled(cron = "0 0 12 * * ?", zone = "Europe/Warsaw")
-    @PostConstruct
-    public void fetchData() throws ConnectException {
-        List<PostEntity> posts = postMapper.mapToEntities(getPosts());
-        if (postRepository.count() == 0) {
-            postRepository.saveAll(posts);
-            log.info("Data inserted");
-        } else {
-            postRepository.updateAllExceptEditedAndDeletedByUser(posts);
-            log.info("Data updated");
-        }
-    }
-
-    private List<PostDto> getPosts() throws ConnectException {
+    public List<PostDto> getPosts() throws ConnectException {
         RestTemplate restTemplate = new RestTemplate();
         ResponseEntity<PostDto[]> responseEntity;
         try {
