+++ src/main/java/pl/apzumi/task/service/PostService.java
@@ -3,6 +3,7 @@ package pl.apzumi.task.service;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pl.apzumi.task.domain.PostEntity;
@@ -12,7 +13,9 @@ import pl.apzumi.task.dto.PostUpdateDto;
 import pl.apzumi.task.mappers.PostMapper;
 import pl.apzumi.task.repository.PostRepository;
 
+import javax.annotation.PostConstruct;
 import javax.persistence.EntityNotFoundException;
+import java.net.ConnectException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -21,8 +24,23 @@ import java.util.List;
 @AllArgsConstructor
 @Transactional
 public class PostService {
+    private final PostFetchService postFetchService;
     private final PostRepository postRepository;
     private final PostMapper postMapper;
+
+    @Scheduled(cron = "0 0 12 * * ?", zone = "Europe/Warsaw")
+    @PostConstruct
+    public void fetchData() throws ConnectException {
+        List<PostDto> postsFromApi = postFetchService.getPosts();
+        List<PostEntity> postEntities = postMapper.mapToEntities(postsFromApi);
+        if (postRepository.count() == 0) {
+            postRepository.saveAll(postEntities);
+            log.info("Data inserted");
+        } else {
+            postRepository.updateAllExceptEditedAndDeletedByUser(postEntities);
+            log.info("Data updated");
+        }
+    }
 
     public List<PostDto> getFilteredPostsByTitle(String title) {
         if (title == null || title.isBlank()) {
@@ -51,5 +69,4 @@ public class PostService {
                 .orElseThrow(() ->
                         new EntityNotFoundException("Resource with id " + id + " not found"));
     }
-
 }
